moatmetrics/requirements.txt
# Core dependencies
python>=3.11

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Data processing
pandas==2.1.4
numpy==1.26.3
duckdb==0.9.2
openpyxl==3.1.2

# Machine Learning & Explainability
scikit-learn==1.3.2
shap==0.44.0
xgboost==2.0.3

# Visualization (for SHAP)
matplotlib==3.8.2

# Report generation
reportlab==4.0.8
jinja2==3.1.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async support
aiofiles==23.2.1

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3.post1
